LayoutTests/inspector/css/modify-rule-selector-expected.txt

Testing that selectors are able to be modified more than once.


Selector before mutation: .foo
Selector after mutation: span.foo

Selector before mutation: span.foo
Selector after mutation: body > .foo
